requirements.txt
sqlalchemy
pyqt
loguru
mypy
psycopg2-binary
pytest
qtmodern
